Imported using git-ubuntu import.

+++ test/shell/profiles-thin.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/usr/bin/env bash
 
 # Copyright (C) 2014 Red Hat, Inc. All rights reserved.
 #
@@ -31,10 +31,11 @@ aux prepare_profiles "thin-performance"
 aux prepare_scsi_debug_dev $DEV_SIZE sector_size=4096 opt_blks=256 || skip
 EXPECT=1048576
 check sysfs "$(< SCSI_DEBUG_DEV)" queue/optimal_io_size "$EXPECT"
-aux prepare_pvs 1 $DEV_SIZE
+aux prepare_pvs 1 "$DEV_SIZE"
 
 # Check we are not running on buggy kernel (broken lcm())
 # If so, turn chunk_size test into  'should'
+SHOULD=""
 check sysfs "$dev1" queue/optimal_io_size "$EXPECT" || SHOULD=should
 
 vgcreate $vg "$dev1"
@@ -62,7 +63,7 @@ vgremove -ff $vg
 
 if test -d "$DM_DEV_DIR/$vg" ; then
 	should not echo "Udev has left \"$DM_DEV_DIR/$vg\"!"
-	rm -rf "$DM_DEV_DIR/$vg"
+	rm -rf "${DM_DEV_DIR:?/dev}/$vg"
 fi
 
 # The profile must be also applied if using the profile
@@ -74,3 +75,5 @@ check vg_field $vg profile "thin-performance"
 check lv_field $vg/pool_performance_inherited profile ""
 $SHOULD check lv_field $vg/pool_performance_inherited chunk_size 1.00m
 check lv_field $vg/pool_performance_inherited zero ""
+
+vgremove -ff $vg
